refactor: Switching dark and light mode in a separate component
